tree trait
