+++ identifier-forest/src/lib.rs
@@ -10,3 +10,8 @@ pub use crate::position::*;
 mod order;
 pub use crate::order::*;
 
+mod tree_trait;
+pub use crate::tree_trait::*;
+
+mod tree_predefined;
+pub use crate::tree_predefined::*;
